pyproject.toml
[project]
name = "pawfreeze"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.92",
    "pynput>=1.8.1",
    "pyobjc-framework-cocoa>=12.1",
    "pyobjc-framework-quartz>=12.1",
    "pytest>=9.0.2",
    "requests>=2.32.5",
    "rumps>=0.4.0",
    "ultralytics>=8.4.14",
]

[project.scripts]
pawfreeze = "PawFreeze.__main__:main"

[tool.pytest.ini_options]
pythonpath = ["PawFreeze"]